+++ packages/react-client/src/callbacks/fetchQuery.ts
@@ -1,6 +1,5 @@
 import type { QraftClientOptions } from '../qraftAPIClient.js';
 import { callQueryClientMethodWithQueryKey } from '../lib/callQueryClientFetchMethod.js';
-import { callQueryClientMethodWithQueryKey } from '../lib/callQueryClientFetchMethod.js';
 import { ServiceOperationFetchQuery } from '../service-operation/ServiceOperationFetchQuery.js';
 
 export const fetchQuery: <