+++ packages/react-client/src/callbacks/setInfiniteQueryData.ts
@@ -3,8 +3,6 @@ import type { QraftClientOptions } from '../qraftAPIClient.js';
 import type { ServiceOperationSetInfiniteQueryData } from '../service-operation/ServiceOperationSetInfiniteQueryData.js';
 import { InfiniteData } from '@tanstack/query-core';
 import { callQueryClientMethodWithQueryKey } from '../lib/callQueryClientMethodWithQueryKey.js';
-import { InfiniteData } from '@tanstack/query-core';
-import { callQueryClientMethodWithQueryKey } from '../lib/callQueryClientMethodWithQueryKey.js';
 
 export function setInfiniteQueryData<TData>(
   qraftOptions: QraftClientOptions,